+++ src/TestTrackingDiagrams.Extensions.SQS/AmazonSQSConfigExtensions.cs
@@ -1,0 +1,20 @@
+using Amazon.SQS;
+
+namespace TestTrackingDiagrams.Extensions.SQS;
+
+public static class AmazonSQSConfigExtensions
+{
+    public static AmazonSQSConfig WithTestTracking(
+        this AmazonSQSConfig config,
+        SqsTrackingMessageHandlerOptions options)
+    {
+        config.HttpClientFactory = new TrackingHttpClientFactory(
+            new SqsTrackingMessageHandler(options));
+        return config;
+    }
+}
+
+internal class TrackingHttpClientFactory(HttpMessageHandler handler) : Amazon.Runtime.HttpClientFactory
+{
+    public override HttpClient CreateHttpClient(Amazon.Runtime.IClientConfig clientConfig) => new(handler);
+}