+++ src/TestTrackingDiagrams.Extensions.SQS/SqsTrackingVerbosity.cs
@@ -1,0 +1,8 @@
+namespace TestTrackingDiagrams.Extensions.SQS;
+
+public enum SqsTrackingVerbosity
+{
+    Raw,
+    Detailed,
+    Summarised
+}